+++ frontend/src/lib/stores/balance.ts
@@ -1,0 +1,5 @@
+import { writable } from "svelte/store";
+
+const balance = writable(0);
+
+export default balance;
